refactor: move icon component
